+++ example/lib/cool_name.dart
@@ -1,3 +1,4 @@
+import 'package:cqrs/cqrs.dart';
 import 'package:cqrs/contracts.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:equatable/equatable.dart';
@@ -8,31 +9,64 @@ part 'cool_name.g.dart';
 
 final _ = EquatableConfig.stringify = true;
 
-/// [TimeOfDay] but with seconds precision
-class Time with EquatableMixin {
-  const Time(this.hour, this.minute, this.second)
-      : assert(hour >= 0 && hour <= 23),
-        assert(minute >= 0 && minute <= 59),
-        assert(second >= 0 && second <= 59);
+class DurationJsonConverter extends JsonConverter<Duration, String> {
+  const DurationJsonConverter();
 
-  factory Time.fromJson(String json) {
-    final chunks = json.split(':');
-    return Time(
-      int.parse(chunks[0]),
-      int.parse(chunks[1]),
-      int.parse(chunks[2]),
-    );
+  Duration fromJson(String json) {
+    final m = RegExp(r'^-?((\d+)\.)?(\d\d):(\d\d):(\d\d)(\.(\d+))?$')
+        .firstMatch(json)!;
+    return Duration(
+          days: int.parse(m[2] ?? '0'),
+          hours: int.parse(m[3]!),
+          minutes: int.parse(m[4]!),
+          seconds: int.parse(m[5]!),
+          microseconds: int.parse(m[7] ?? '0') ~/ 10,
+        ) *
+        (json.startsWith('-') ? -1 : 1);
   }
 
-  final int hour;
+  String toJson(Duration value) {
+    if (value.isNegative) return '-${toJson(-value)}';
 
-  final int minute;
+    return '${value.inDays}.'
+        '${value.inHours % Duration.hoursPerDay}:'
+        '${value.inMinutes % Duration.minutesPerHour}:'
+        '${value.inSeconds % Duration.secondsPerMinute}.'
+        '${(value.inMicroseconds % Duration.microsecondsPerSecond).toString().padLeft(6, '0')}';
+  }
+}
 
-  final int second;
+class NullableDurationJsonConverter extends JsonConverter<Duration?, String?> {
+  const NullableDurationJsonConverter();
 
-  get props => [hour, minute, second];
+  Duration? fromJson(String? json) =>
+      json == null ? null : const DurationJsonConverter().fromJson(json);
+  String? toJson(Duration? value) =>
+      value == null ? null : const DurationJsonConverter().toJson(value);
+}
 
-  String toJson() => '$hour:$minute:$second';
+/// [TimeOfDay] but with microseconds precision
+class TimeOnly with EquatableMixin {
+  TimeOnly(this._source)
+      : assert(_source < const Duration(days: 1) && !_source.isNegative);
+
+  TimeOnly.fromJson(String json)
+      : this(const DurationJsonConverter().fromJson(json));
+
+  final Duration _source;
+
+  int get hour => _source.inHours % Duration.hoursPerDay;
+
+  int get minute => _source.inMinutes % Duration.minutesPerHour;
+
+  int get second => _source.inSeconds % Duration.secondsPerMinute;
+
+  int get microsecond =>
+      _source.inMicroseconds % Duration.microsecondsPerSecond;
+
+  get props => [_source];
+
+  String toJson() => const DurationJsonConverter().toJson(_source);
 }
 
 @JsonSerializable(fieldRename: FieldRename.pascal)