+++ lib/src/statements/dto_handler.dart
@@ -2,6 +2,7 @@ import 'package:code_builder/code_builder.dart';
 
 import '../attributes/attribute_creator.dart';
 import '../generator_database.dart';
+import '../json_converters/json_converters.dart';
 import '../types/type_creator.dart';
 import '../values/value_creator.dart';
 import 'statement_handler.dart';
@@ -11,8 +12,9 @@ class DtoHandler extends StatementHandler {
     TypeCreator typeCreator,
     ValueCreator valueCreator,
     AttributeCreator attributeCreator,
+    JsonConverters jsonConverters,
     GeneratorDatabase db,
-  ) : super(typeCreator, valueCreator, attributeCreator, db);
+  ) : super(typeCreator, valueCreator, attributeCreator, jsonConverters, db);
 
   @override
   Spec build(Statement statement) {