+++ lib/src/statements/query_handler.dart
@@ -2,6 +2,7 @@ import 'package:code_builder/code_builder.dart';
 
 import '../attributes/attribute_creator.dart';
 import '../generator_database.dart';
+import '../json_converters/json_converters.dart';
 import '../types/type_creator.dart';
 import '../values/value_creator.dart';
 import 'result_factory/internal_result_factory_handler.dart';
@@ -15,15 +16,16 @@ class QueryHandler extends StatementHandler {
     TypeCreator typeCreator,
     ValueCreator valueCreator,
     AttributeCreator attributeCreator,
+    JsonConverters jsonConverters,
     GeneratorDatabase db,
   )   : resultFactoryCreator = ResultFactoryCreator(
           [
             InternalResultFactoryHandler(db),
-            const KnownResultFactoryHandler(),
+            KnownResultFactoryHandler(jsonConverters)
           ],
           typeCreator,
         ),
-        super(typeCreator, valueCreator, attributeCreator, db);
+        super(typeCreator, valueCreator, attributeCreator, jsonConverters, db);
 
   final ResultFactoryCreator resultFactoryCreator;