+++ lib/src/statements/statement_handler.dart
@@ -4,6 +4,7 @@ import 'package:source_helper/source_helper.dart';
 
 import '../attributes/attribute_creator.dart';
 import '../generator_database.dart';
+import '../json_converters/json_converters.dart';
 import '../proto/contracts.pb.dart';
 import '../types/type_creator.dart';
 import '../utils/rename_field.dart';
@@ -18,12 +19,14 @@ abstract class StatementHandler {
     this.typeCreator,
     this.valueCreator,
     this.attributeCreator,
+    this.jsonConverters,
     this.db,
   );
 
   final TypeCreator typeCreator;
   final ValueCreator valueCreator;
   final AttributeCreator attributeCreator;
+  final JsonConverters jsonConverters;
   final GeneratorDatabase db;
 
   Spec build(Statement statement);
@@ -52,6 +55,11 @@ abstract class StatementHandler {
     final genericFactories = typeDescriptor.genericParameters
         .map((e) => _GenericFactory(e.name))
         .toList();
+
+    final neededConverters = typeDescriptor.properties
+        .map((p) => jsonConverters.getConverter(p.type))
+        .whereType<Class>()
+        .toSet();
 
     return Class((b) {
       b
@@ -122,13 +130,15 @@ abstract class StatementHandler {
         ..types.addAll(
           typeDescriptor.genericParameters.map((t) => refer(t.name)),
         )
-        ..annotations.add(
+        ..annotations.addAll([
           CodeExpression(
             Code(
               'JsonSerializable(fieldRename: FieldRename.pascal${genericFactories.isEmpty ? '' : ', genericArgumentFactories: true'})',
             ),
           ),
-        )
+          for (final converter in neededConverters)
+            CodeExpression(Code('${converter.name}()')),
+        ])
         ..docs.addAll([
           ...toDartdoc(statement.comment),
           ...statement.attributes.map(attributeCreator.create),