+++ lib/src/time_class.dart
@@ -1,23 +1,43 @@
 import 'package:code_builder/code_builder.dart';
 
-const timeClassName = 'Time';
+import 'json_converters/duration_json_converter.dart';
 
-const _timeClassFields = ['hour', 'minute', 'second'];
+const timeClassName = 'TimeOnly';
 
 final timeClass = Class(
   (b) => b
     ..name = timeClassName
     ..fields.addAll([
-      for (final name in _timeClassFields)
-        Field(
-          (b) => b
-            ..name = name
-            ..modifier = FieldModifier.final$
-            ..type = refer('int'),
-        ),
       Field(
         (b) => b
-          ..name = 'get props => [${_timeClassFields.join(',')}]'
+          ..name = '_source'
+          ..type = refer('Duration')
+          ..modifier = FieldModifier.final$,
+      ),
+      Field(
+        (b) => b
+          ..name = 'get hour => _source.inHours % Duration.hoursPerDay'
+          ..type = refer('int'),
+      ),
+      Field(
+        (b) => b
+          ..name = 'get minute => _source.inMinutes % Duration.minutesPerHour'
+          ..type = refer('int'),
+      ),
+      Field(
+        (b) => b
+          ..name = 'get second => _source.inSeconds % Duration.secondsPerMinute'
+          ..type = refer('int'),
+      ),
+      Field(
+        (b) => b
+          ..name =
+              'get microsecond => _source.inMicroseconds % Duration.microsecondsPerSecond'
+          ..type = refer('int'),
+      ),
+      Field(
+        (b) => b
+          ..name = 'get props => [_source]'
           ..type = refer(''),
       ),
     ])
@@ -26,23 +46,20 @@ final timeClass = Class(
       Constructor(
         (b) => b
           ..requiredParameters.addAll([
-            for (final name in _timeClassFields)
-              Parameter(
-                (b) => b
-                  ..name = name
-                  ..toThis = true,
-              ),
+            Parameter(
+              (b) => b
+                ..name = '_source'
+                ..toThis = true,
+            ),
           ])
-          ..initializers.addAll(const [
-            Code('assert(hour >= 0 && hour <= 23)'),
-            Code('assert(minute >= 0 && minute <= 59)'),
-            Code('assert(second >= 0 && second <= 59)'),
-          ])
-          ..constant = true,
+          ..initializers.add(
+            const Code(
+              'assert(_source < const Duration(days: 1) && !_source.isNegative)',
+            ),
+          ),
       ),
       Constructor(
         (b) => b
-          ..factory = true
           ..name = 'fromJson'
           ..requiredParameters.add(
             Parameter(
@@ -51,11 +68,8 @@ final timeClass = Class(
                 ..type = refer('String'),
             ),
           )
-          ..body = const Code(
-            """
-            final chunks = json.split(':');
-            return $timeClassName(int.parse(chunks[0]), int.parse(chunks[1]), int.parse(chunks[2]),);
-            """,
+          ..initializers.add(
+            Code('this(const ${durationJsonConverter.name}().fromJson(json))'),
           ),
       )
     ])
@@ -65,8 +79,9 @@ final timeClass = Class(
           ..name = 'toJson'
           ..lambda = true
           ..returns = refer('String')
-          ..body = literalString(r'$hour:$minute:$second').code,
+          ..body =
+              Code('const ${durationJsonConverter.name}().toJson(_source)'),
       ),
     )
-    ..docs.add('/// [TimeOfDay] but with seconds precision'),
+    ..docs.add('/// [TimeOfDay] but with microseconds precision'),
 );