+++ lib/src/types/known_type_handler.dart
@@ -4,43 +4,11 @@ import 'type_handler.dart';
 class KnownTypeHandler extends TypeHandler {
   const KnownTypeHandler();
 
-  static const knownTypeToDartType = {
-    KnownType.Object_: 'Object',
-    KnownType.String: 'String',
-    KnownType.Guid: 'String',
-    KnownType.Uri: 'Uri',
-    KnownType.Boolean: 'bool',
-    KnownType.UInt8: 'int',
-    KnownType.Int8: 'int',
-    KnownType.Int16: 'int',
-    KnownType.UInt16: 'int',
-    KnownType.Int32: 'int',
-    KnownType.UInt32: 'int',
-    KnownType.Int64: 'int',
-    KnownType.UInt64: 'int',
-    KnownType.Float: 'double',
-    KnownType.Double: 'double',
-    KnownType.Decimal: 'double',
-    KnownType.Date: 'DateTime',
-    KnownType.Time: timeClassName,
-    KnownType.TimeSpan: 'Duration',
-    KnownType.DateTime: 'DateTime',
-    KnownType.DateTimeOffset: 'DateTime',
-    KnownType.Array: 'List',
-    KnownType.Map: 'Map',
-    KnownType.Query: 'IRemoteQuery',
-    KnownType.Command: 'IRemoteCommand',
-    KnownType.AuthorizeWhenAttribute: 'TODO',
-    KnownType.AuthorizeWhenHasAnyOfAttribute: 'TODO',
-    KnownType.QueryCacheAttribute: 'TODO',
-    KnownType.Attribute: 'TODO',
-  };
-
   @override
   String build(TypeRef typeRef, Iterable<TypeHandler> handlers) {
     assert(canHandle(typeRef));
 
-    final name = _toDartType(typeRef.known.type);
+    final name = toDartType(typeRef.known.type);
 
     if (typeRef.known.arguments.isNotEmpty) {
       final argumentList = typeRef.known.arguments
@@ -53,16 +21,60 @@ class KnownTypeHandler extends TypeHandler {
     return name;
   }
 
-  String _toDartType(KnownType knownType) {
-    final dartTypeStr = knownTypeToDartType[knownType];
-
-    if (dartTypeStr == null) {
-      throw UnimplementedError(
-        'There is a missing mapping for the type $knownType',
-      );
+  static String toDartType(KnownType knownType) {
+    switch (knownType) {
+      case KnownType.Object_:
+        return 'Object';
+      case KnownType.String:
+      case KnownType.Guid:
+        return 'String';
+      case KnownType.Uri:
+        return 'Uri';
+      case KnownType.Boolean:
+        return 'bool';
+      case KnownType.UInt8:
+      case KnownType.Int8:
+      case KnownType.Int16:
+      case KnownType.UInt16:
+      case KnownType.Int32:
+      case KnownType.UInt32:
+      case KnownType.Int64:
+      case KnownType.UInt64:
+        return 'int';
+      case KnownType.Float:
+      case KnownType.Double:
+      case KnownType.Decimal:
+        return 'double';
+      case KnownType.TimeOnly:
+      case KnownType.Time:
+        return timeClassName;
+      case KnownType.TimeSpan:
+        return 'Duration';
+      case KnownType.Date:
+      case KnownType.DateTime:
+      case KnownType.DateTimeOffset:
+      case KnownType.DateOnly:
+        return 'DateTime';
+      case KnownType.Array:
+        return 'List';
+      case KnownType.Map:
+        return 'Map';
+      case KnownType.Query:
+        return 'IRemoteQuery';
+      case KnownType.Command:
+        return 'IRemoteCommand';
+      case KnownType.CommandResult:
+        return 'CommandResult';
+      case KnownType.AuthorizeWhenAttribute:
+      case KnownType.AuthorizeWhenHasAnyOfAttribute:
+      case KnownType.QueryCacheAttribute:
+      case KnownType.Attribute:
+        throw UnimplementedError(
+          'There is a missing mapping for the type $knownType',
+        );
     }
 
-    return dartTypeStr;
+    throw StateError('unreachable');
   }
 
   @override