+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Unreleased
+
+- **Breaking**: require .NET 6
+- Add support for `CommandResult`
+- Add support for `TimeOnly` and `DateOnly`
+- Correctly deserialize `TimeSpan`
+
 # 0.2.4
 
 - Relax `json_serializable` version requirement (now `4.x.x`, `5.x.x`, and `6.x.x` are accepted)